Refactor dev command
