+++ pkg/kev/types.go
@@ -18,6 +18,7 @@ package kev
 
 import (
 	composego "github.com/compose-spec/compose-go/types"
+	"github.com/spf13/cobra"
 )
 
 // Manifest contains the tracked project's docker-compose sources and deployment environments
@@ -101,3 +102,12 @@ type change struct {
 	Target string
 	Index  interface{}
 }
+
+// ErrorHandler is a callback function that handles error and returns error
+type ErrorHandler func(error) error
+
+// ChangeHandler is a callback function that handles change and returns error, e.g. change event when in dev mode
+type ChangeHandler func(string) error
+
+// PreRunCmd is a command expected to run before the prior to current command
+type PreRunCmd func(*cobra.Command, []string) error
